Moved TpaCommand to [...].command.tpa
